# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IMOK Api",
        "description": "This API is meant to be used along with the IMOK_Mobile and IMOK_Desktop apps. All routes require you to be authenticated, so your first step is to send your credentials to api/auth/login and get a bearer token.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/customers/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get a customer",
                "operationId": "App\\Http\\Controllers\\CustomerController::getOneById",
                "responses": {
                    "404": {
                        "description": "Customer does not exist"
                    },
                    "200": {
                        "description": "Customer",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update a customer",
                "operationId": "App\\Http\\Controllers\\CustomerController::update",
                "responses": {
                    "404": {
                        "description": "Customer does not exist"
                    },
                    "409": {
                        "description": "Customer could not be updated"
                    },
                    "201": {
                        "description": "Customer updated"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/customers/find/{argument}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get a customer where argument can be found in name or mail",
                "operationId": "App\\Http\\Controllers\\CustomerController::getWhere",
                "responses": {
                    "404": {
                        "description": "No customer found"
                    },
                    "200": {
                        "description": "Matching customer(s)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Customer"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/estates": {
            "get": {
                "tags": [
                    "Estates"
                ],
                "summary": "Get all estate",
                "operationId": "App\\Http\\Controllers\\EstateController::getAll",
                "responses": {
                    "200": {
                        "description": "Estates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Estate"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Estates"
                ],
                "summary": "Create estate",
                "operationId": "App\\Http\\Controllers\\EstateController::create",
                "requestBody": {
                    "description": "Some arguments are required",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "customers_id": {
                                        "description": "CREATE A NEW ESTATE",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estate created"
                    },
                    "409": {
                        "description": "Estate could not be created"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/estates/{id}": {
            "get": {
                "tags": [
                    "Estates"
                ],
                "summary": "Get estate",
                "operationId": "App\\Http\\Controllers\\EstateController::getOneById",
                "responses": {
                    "200": {
                        "description": "Estate",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Estate"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Estate not found"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Estates"
                ],
                "summary": "Update an estate",
                "operationId": "App\\Http\\Controllers\\EstateController::update",
                "responses": {
                    "201": {
                        "description": "Estate updated"
                    },
                    "404": {
                        "description": "Estate not found"
                    },
                    "409": {
                        "description": "Estate could not be updated"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/estates/search": {
            "post": {
                "tags": [
                    "Estates"
                ],
                "summary": "Search estate with matching argument",
                "operationId": "App\\Http\\Controllers\\EstateController::getWhere",
                "requestBody": {
                    "description": "Pass any of these arguments",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "city": {
                                        "description": "GET AN ESTATE WHERE LIKE IN ATTR",
                                        "type": "string",
                                        "example": "amiens"
                                    },
                                    "minBudget": {
                                        "description": "GET AN ESTATE WHERE LIKE IN ATTR",
                                        "type": "string",
                                        "example": "10000"
                                    },
                                    "maxBudget": {
                                        "description": "GET AN ESTATE WHERE LIKE IN ATTR",
                                        "type": "string",
                                        "example": "999999"
                                    },
                                    "minSize": {
                                        "description": "GET AN ESTATE WHERE LIKE IN ATTR",
                                        "type": "string",
                                        "example": "30"
                                    },
                                    "maxSize": {
                                        "description": "GET AN ESTATE WHERE LIKE IN ATTR",
                                        "type": "string",
                                        "example": "300"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Matching Estates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Estate"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No matching Estate found"
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Customer": {
                "title": "Customer",
                "properties": {
                    "firstname": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "lastname": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "street": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "complement": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "mail": {
                        "description": "Class Customer",
                        "type": "email"
                    },
                    "id_marital_status": {
                        "description": "Class Customer",
                        "type": "integer"
                    },
                    "id_cities": {
                        "description": "Class Customer",
                        "type": "integer"
                    },
                    "civility": {
                        "description": "Class Customer",
                        "type": "integer"
                    },
                    "birthdate": {
                        "description": "Class Customer",
                        "type": "string",
                        "format": "date"
                    },
                    "date_register": {
                        "description": "Class Customer",
                        "type": "string",
                        "format": "date"
                    },
                    "zip_code": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "city": {
                        "description": "Class Customer",
                        "type": "string"
                    },
                    "marital_status": {
                        "description": "Class Customer",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Estate": {
                "title": "Estate",
                "properties": {
                    "id": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "rooms": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "facilities": {
                        "description": "Class Estate",
                        "type": "object"
                    },
                    "street": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "complement": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "description": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "room_numbers": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "bedroom_numbers": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "joint-ownership": {
                        "description": "Class Estate",
                        "type": "boolean"
                    },
                    "annual_fees": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "double"
                    },
                    "price": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "double"
                    },
                    "rent": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "double"
                    },
                    "condominium": {
                        "description": "Class Estate",
                        "type": "boolean"
                    },
                    "condominium_fees": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "double"
                    },
                    "property_tax": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "double"
                    },
                    "housing_tax": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "double"
                    },
                    "energy_consumption": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "double"
                    },
                    "gas_emission": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "double"
                    },
                    "size": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "double"
                    },
                    "carrez_size": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "double"
                    },
                    "floor": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "floor_number": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "renovation": {
                        "description": "Class Estate",
                        "type": "boolean"
                    },
                    "id_customers": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "id_build_Dates": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "id_outside_conditions": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "id_heating_types": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "id_districts": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "id_expositions": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "id_cities": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "id_estate_types": {
                        "description": "Class Estate",
                        "type": "integer"
                    },
                    "period": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "zip_code": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "city": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "city_lat": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "float"
                    },
                    "cities_lng": {
                        "description": "Class Estate",
                        "type": "number",
                        "format": "float"
                    },
                    "owner_firsname": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "owner_lastname": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "exposition": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "heating_type": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "notes": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "outside_condition": {
                        "description": "Class Estate",
                        "type": "string"
                    },
                    "estate_type": {
                        "description": "Class Estate",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "JWT Auth Token",
                "name": "JWT",
                "in": "header"
            }
        }
    }
}